+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add createInventoryItem function to handle item creation in Airtable
+- Add saving state management to QuickAddForm and AddInventoryItemForm
 
 ---